requirements.txt
# torch==2.3
# numpy==1.26.4
# pillow==10.3.0
# fire==0.6.0
# transformers==4.41.2
# safetensors==0.4.3


torch
numpy
pillow
fire
transformers
safetensors
